Improving error cases.

+++ pypdftk.py
@@ -15,12 +15,10 @@ import shutil
 PDFTK_PATH = '/usr/bin/pdftk'
 
 
-def run_command(command):
+def run_command(command, shell=False):
     ''' run a system command and yield output '''
-    p = subprocess.Popen(command,
-                         stdout=subprocess.PIPE,
-                         stderr=subprocess.STDOUT)
-    return iter(p.stdout.readline, b'')
+    p = subprocess.check_output(command, shell=shell)
+    return p.split('\n')
 
 
 def get_num_pages(pdf_path):
@@ -36,24 +34,21 @@ def fill_form(pdf_path, datas={}, out_file=None, flatten=True):
         Fills a PDF form with given dict input data.
         Return temp file if no out_file provided.
     '''
-    error = False
+    cleanOnFail = False
     tmp_fdf = gen_xfdf(datas)
     if not out_file:
+        cleanOnFail = True
         handle, out_file = tempfile.mkstemp()
 
     cmd = "%s %s fill_form %s output %s" % (PDFTK_PATH, pdf_path, tmp_fdf, out_file)
     if flatten:
         cmd += ' flatten'
     try:
-        p = subprocess.Popen(cmd, shell=True)
+        run_command(cmd, True)
     except:
-        error = True
-    p.wait()
-    os.remove(tmp_fdf)
-    if error or not os.path.isfile(out_file):
-        print 'Error filling PDF page %s' % pdf_path
-        # if merge failed, return the original PDF
-        shutil.copyfile(pdf_path, out_file)
+        if cleanOnFail:
+            os.remove(tmp_fdf)
+        raise
     return out_file
 
 
@@ -62,14 +57,21 @@ def concat(files, out_file=None):
         Merge multiples PDF files.
         Return temp file if no out_file provided.
     '''
+    cleanOnFail = False
     if not out_file:
+        cleanOnFail = True
         handle, out_file = tempfile.mkstemp()
     if len(files) == 1:
         shutil.copyfile(files[0], out_file)
     args = [PDFTK_PATH]
     args += files
     args += ['cat', 'output', out_file]
-    run_command(args)
+    try:
+        run_command(args)
+    except:
+        if cleanOnFail:
+            os.remove(out_file)
+        raise
     return out_file
 
 
@@ -78,10 +80,17 @@ def split(pdf_path, out_dir=None):
         Split a single PDF file into pages.
         Use a temp directory if no out_dir provided.
     '''
+    cleanOnFail = False
     if not out_dir:
+        cleanOnFail = True
         out_dir = tempfile.mkdtemp()
     out_pattern = '%s/page_%%02d.pdf' % out_dir
-    run_command((PDFTK_PATH, pdf_path, 'burst', 'output', out_pattern))
+    try:
+        run_command((PDFTK_PATH, pdf_path, 'burst', 'output', out_pattern))
+    except:
+        if cleanOnFail:
+            shutil.rmtree(out_dir)
+        raise
     out_files = os.listdir(out_dir)
     out_files.sort()
     return [os.path.join(out_dir, filename) for filename in out_files]
